% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_symbol}
\alias{assert_symbol}
\title{Is an Argument a Symbol?}
\usage{
assert_symbol(
  arg,
  optional = FALSE,
  arg_name = gsub("^enexpr\\\\((.*)\\\\)$", "\\\\1", rlang::caller_arg(arg)),
  message = NULL,
  class = "assert_symbol",
  call = parent.frame()
)
}
\arguments{
\item{arg}{A function argument to be checked. Must be a \code{symbol}. See examples.

\describe{
\item{Default value}{none}
}}

\item{optional}{Is the checked argument optional? If set to \code{FALSE} and \code{arg}
is \code{NULL} then an error is thrown.

\describe{
\item{Default value}{\code{FALSE}}
}}

\item{arg_name}{By default the expression specified for \code{arg} is used. If it is
of the form \verb{enexpr(<argument name>)}, the \code{enexpr()} part is removed. For example
if \code{arg = enexpr(filter_add)} is specified, \code{arg_name} defaults to \code{"filter_add"}

\describe{
\item{Default value}{\code{gsub("^enexpr\\((.*)\\)$", "\\1", rlang::caller_arg(arg))}}
}}

\item{message}{string passed to \code{cli::cli_abort(message)}.
When \code{NULL}, default messaging is used (see examples for default messages).
\code{"{arg_name}"} can be used in messaging.

\describe{
\item{Default value}{\code{NULL}}
}}

\item{class}{Subclass of the condition.}

\item{call}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}
}
\value{
The function throws an error if \code{arg} is not a symbol and returns the input
invisibly otherwise.
}
\description{
Checks if an argument is a symbol
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(rlang)
dm <- dplyr::tribble(
  ~DOMAIN,      ~USUBJID,
  "DM",    "01-701-1015",
  "DM",    "01-701-1016",
)
example_fun <- function(dat, var) {
  var <- assert_symbol(enexpr(var))
  select(dat, !!var)
}

example_fun(dm, USUBJID)

try(example_fun(dm))

try(example_fun(dm, "USUBJID"))

try(example_fun(dm, toupper(PARAMCD)))
}
\seealso{
Checks for valid input and returns warning or errors messages:
\code{\link[=assert_atomic_vector]{assert_atomic_vector()}},
\code{\link[=assert_character_scalar]{assert_character_scalar()}},
\code{\link[=assert_character_vector]{assert_character_vector()}},
\code{\link[=assert_data_frame]{assert_data_frame()}},
\code{\link[=assert_date_vector]{assert_date_vector()}},
\code{\link[=assert_expr]{assert_expr()}},
\code{\link[=assert_expr_list]{assert_expr_list()}},
\code{\link[=assert_filter_cond]{assert_filter_cond()}},
\code{\link[=assert_function]{assert_function()}},
\code{\link[=assert_integer_scalar]{assert_integer_scalar()}},
\code{\link[=assert_list_element]{assert_list_element()}},
\code{\link[=assert_list_of]{assert_list_of()}},
\code{\link[=assert_logical_scalar]{assert_logical_scalar()}},
\code{\link[=assert_named]{assert_named()}},
\code{\link[=assert_numeric_vector]{assert_numeric_vector()}},
\code{\link[=assert_one_to_one]{assert_one_to_one()}},
\code{\link[=assert_param_does_not_exist]{assert_param_does_not_exist()}},
\code{\link[=assert_s3_class]{assert_s3_class()}},
\code{\link[=assert_same_type]{assert_same_type()}},
\code{\link[=assert_unit]{assert_unit()}},
\code{\link[=assert_vars]{assert_vars()}},
\code{\link[=assert_varval_list]{assert_varval_list()}}
}
\concept{assertion}
\keyword{assertion}
