% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters.R
\name{date_convention_linter}
\alias{date_convention_linter}
\title{Linter object creator for *DTC/--DTM custom linter}
\usage{
date_convention_linter(source_file)
}
\description{
This function creates the \code{Linter} object corresponding to
\code{date_convention_linter_source()}.
}
\keyword{internal}
