% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters.R
\name{date_convention_linter_source}
\alias{date_convention_linter_source}
\title{Custom linter for *DTC/--DTM}
\usage{
date_convention_linter_source(source_file)
}
\arguments{
\item{source_file}{source_file to check

\describe{
\item{Default value}{none}
}}
}
\description{
This linter function checks whether a file contains a comment with "*DTC"
or "--DTM" anywhere in the file, including inline.
}
\keyword{internal}
