% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_fun.R
\name{demo_fun}
\alias{demo_fun}
\title{A Demo Function}
\usage{
demo_fun(x, number = 1, letter = "a")
}
\arguments{
\item{x}{An argument}

\item{number}{A number}

\item{letter}{A letter}
}
\description{
This function is used to demonstrate the custom tags of the `rdx_roclet()`.
}
\keyword{internal}
