% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\name{dquote}
\alias{dquote}
\title{Wrap a String in Double Quotes}
\usage{
dquote(x)
}
\arguments{
\item{x}{A character vector

\describe{
\item{Default value}{none}
}}
}
\value{
If the input is \code{NULL}, the text \code{"NULL"} is returned. Otherwise, the
input in double quotes is returned.
}
\description{
Wrap a string in double quotes, e.g., for displaying character values in
messages.
}
\seealso{
Helpers for working with Quotes and Quoting:
\code{\link[=backquote]{backquote()}},
\code{\link[=squote]{squote()}}
}
\concept{quote}
\keyword{quote}
