% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_rdx.R
\name{execute_example}
\alias{execute_example}
\title{Execute Example Code}
\usage{
execute_example(code, expected_cnds = NULL, env = caller_env())
}
\arguments{
\item{code}{The example code

The code is expected to be a character vector of R code lines.

\describe{
\item{Permitted values}{A character vector}
\item{Default value}{none}
}}

\item{expected_cnds}{Expected conditions

\describe{
\item{Permitted values}{A character vector}
\item{Default value}{\code{NULL}}
}}

\item{env}{The environment in which to evaluate the example code

\describe{
\item{Permitted values}{An environment}
\item{Default value}{\code{caller_env()}}
}}
}
\value{
A character vector of the input code and the output and messages
created by the code. Output and messages are prefixed by \code{"#>"}.
}
\description{
This function executes the example code and captures the output and messages.
If the example code issues an unexpected message, an error is issued.
}
\keyword{internal}
\section{Examples}{
\subsection{Execute Example Code}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{suppressWarnings(admiraldev:::execute_example("1 + 1"))
#> [1] "1 + 1\n#> [1] 2"

suppressWarnings(admiraldev:::execute_example("log(-1)"))
#> [1] "log(-1)\n#> [1] NaN"

suppressWarnings(admiraldev:::execute_example("log(-1)", expected_cnds = "warning"))
#> [1] "log(-1)\n#> [1] NaN"}\if{html}{\out{</div>}}}}

