% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quo.R
\name{expr_c}
\alias{expr_c}
\title{Concatenate One or More Expressions}
\usage{
expr_c(...)
}
\arguments{
\item{...}{One or more expressions or list of expressions

\describe{
\item{Default value}{none}
}}
}
\value{
A list of expressions
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please use \code{c()} instead.
}
\seealso{
Other deprecated:
\code{\link{\%or\%}},
\code{\link[=arg_name]{arg_name()}},
\code{\link[=enumerate]{enumerate()}},
\code{\link[=friendly_type_of]{friendly_type_of()}},
\code{\link[=valid_time_units]{valid_time_units()}},
\code{\link[=what_is_it]{what_is_it()}}
}
\concept{deprecated}
\keyword{deprecated}
