% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{filter_if}
\alias{filter_if}
\title{Optional Filter}
\usage{
filter_if(dataset, filter)
}
\arguments{
\item{dataset}{Input dataset

\describe{
\item{Default value}{none}
}}

\item{filter}{A filter condition. Must be an expression.

\describe{
\item{Default value}{none}
}}
}
\value{
A \code{data.frame} containing all rows in \code{dataset} matching \code{filter} or
just \code{dataset} if \code{filter} is \code{NULL}
}
\description{
Filters the input dataset if the provided expression is not \code{NULL}
}
\seealso{
Developer Utility Functions:
\code{\link{\%notin\%}},
\code{\link[=contains_vars]{contains_vars()}},
\code{\link[=convert_dtm_to_dtc]{convert_dtm_to_dtc()}},
\code{\link[=extract_vars]{extract_vars()}},
\code{\link[=vars2chr]{vars2chr()}}
}
\concept{dev_utility}
\keyword{dev_utility}
