% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_friendly_type.R
\name{friendly_type_of}
\alias{friendly_type_of}
\title{Return English-friendly messaging for object-types}
\usage{
friendly_type_of(x, value = TRUE, length = FALSE)
}
\arguments{
\item{x}{Any R object.

\describe{
\item{Default value}{none}
}}

\item{value}{Whether to describe the value of \code{x}.

\describe{
\item{Default value}{\code{TRUE}}
}}

\item{length}{Whether to mention the length of vectors and lists.

\describe{
\item{Default value}{\code{FALSE}}
}}
}
\value{
A string describing the type. Starts with an indefinite
article, e.g. "an integer vector".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please use \code{cli} functionality instead.
}
\details{
This helper function aids us in forming user-friendly messages that gets
called through \code{what_is_it()}, which is often used in the assertion functions
to identify what object-type the user passed through an argument instead of
an expected-type.
}
\seealso{
Other deprecated:
\code{\link{\%or\%}},
\code{\link[=arg_name]{arg_name()}},
\code{\link[=enumerate]{enumerate()}},
\code{\link[=expr_c]{expr_c()}},
\code{\link[=valid_time_units]{valid_time_units()}},
\code{\link[=what_is_it]{what_is_it()}}
}
\concept{deprecated}
\keyword{deprecated}
