% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_duplicates}
\alias{get_duplicates}
\title{Get Duplicates From a Vector}
\usage{
get_duplicates(x)
}
\arguments{
\item{x}{An atomic vector

\describe{
\item{Default value}{none}
}}
}
\value{
A vector of the same type as \code{x} contain duplicate values
}
\description{
Get Duplicates From a Vector
}
\examples{
get_duplicates(1:10)

get_duplicates(c("a", "a", "b", "c", "d", "d"))
}
\seealso{
Brings something to you!?!:
\code{\link[=get_constant_vars]{get_constant_vars()}},
\code{\link[=get_dataset]{get_dataset()}},
\code{\link[=get_source_vars]{get_source_vars()}}
}
\concept{get}
\keyword{get}
