% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_rdx.R
\name{get_param_tag}
\alias{get_param_tag}
\title{Get the content of the `@param` tag}
\usage{
get_param_tag(act_param, defaults)
}
\arguments{
\item{act_param}{The content of the argument description to format

  A list with three named elements is expected:

  - `tag`: The `@param` tag
  - `permitted`: The permitted values
  - `default`: The default value}

\item{defaults}{A list of default values from the function call}
}
\value{
The formatted `@param` tag
}
\description{
This function adds the permitted values and the default value to the
description of the `@param` tag. They are added as a definition list
(`\describe{}`).
}
\keyword{internal}
