% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_valid_dtc}
\alias{is_valid_dtc}
\title{Is this string a valid DTC?}
\usage{
is_valid_dtc(arg, check_dtc = FALSE)
}
\arguments{
\item{arg}{The string to check.

\describe{
\item{Permitted values}{a character scalar, i.e., a character vector of length one}
\item{Default value}{none}
}}

\item{check_dtc}{If \code{TRUE}, the function will check if the string is a real,
date or datetime (e.g. "2020-02-31" is not real). Note that \code{NA} is a valid
datetime.

\describe{
\item{Permitted values}{\code{TRUE}, \code{FALSE}}
\item{Default value}{\code{FALSE}}
}}
}
\value{
\code{TRUE} if the argument is a valid \code{--DTC} string, \code{FALSE} otherwise
}
\description{
Is this string a valid DTC?
}
\details{
If \code{check_dtc} is \code{FALSE}, the function only checks if the format of the
string is valid, i.e. it will not check if the actual date/datetime is real, so
calls such as \verb{is_valid_dtc("2020-13", check_dtc = FALSE))} and
\code{is_valid_dtc("2020-12-01T25:00:00", check_dtc = FALSE)} will return\code{TRUE}.
}
\seealso{
Identifies type of Object with return of TRUE/FALSE:
\code{\link[=is_auto]{is_auto()}},
\code{\link[=is_order_vars]{is_order_vars()}}
}
\concept{is}
\keyword{is}
