% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{anti_join}
\alias{anti_join}
\alias{inner_join}
\alias{left_join}
\title{Join Functions}
\usage{
anti_join(x, y, by = NULL, copy = FALSE, ...)

inner_join(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

left_join(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)
}
\arguments{
\item{x}{\code{data.frame}

\describe{
\item{Default value}{none}
}}

\item{y}{\code{data.frame}

\describe{
\item{Default value}{none}
}}

\item{by}{\code{character} vector

\describe{
\item{Default value}{\code{NULL}}
}}

\item{copy}{\code{logical}

\describe{
\item{Default value}{\code{FALSE}}
}}

\item{...}{Additional arguments

\describe{
\item{Default value}{none}
}}

\item{suffix}{\code{character} vector

\describe{
\item{Default value}{\code{NULL}}
}}
}
\value{
\code{data.frame}
}
\description{
The \verb{*_join()} functions from \code{{dplyr}} without a warning on different attributes
in datasets.
}
\keyword{joins}
