% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_rdx.R
\name{parse_code}
\alias{parse_code}
\title{Parse Code}
\usage{
parse_code(code)
}
\arguments{
\item{code}{The code to parse}
}
\value{
A list of expressions and source references

Each item of the list is a list with the following elements:
- `expr`: The expression
- `srcref`: The source reference
- `eval`: A logical indicating whether the expression should be evaluated,
  i.e., it is not a comment or an empty line.
}
\description{
The function parses the code and returns a list of expressions and source
references.
}
\examples{
parse_code("1+1\n2*2")

parse_code(c("# sum:", "sum(\n  1, #first\n  2\n)"))
}
\keyword{documentation}
