% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_rdx.R
\name{rdx_roclet}
\alias{rdx_roclet}
\title{Roclet Extending the Standard rd Roclet}
\usage{
rdx_roclet()
}
\description{
This roclet extends the standard `rd` roclet by allowing
- to add permitted values and default values to the `@param` tag and
- to add a caption and a description to examples.
}
\details{
The following tags are supported:

- `@permitted`: Permitted values for the argument. Permitted value description
    which are used for several arguments/functions can be stored in
    `inst/roxygen/rdx_meta.R`. For example:
    ```
    list(
      rdx_permitted_values = list(
        mode = "`\"first\"`, `\"last\"`",
        msg_type = "`\"none\"`, `\"message\"`, `\"warning\"`, `\"error\"`"
      )
    )
    ```
    The reference to the permitted values is done by specifying the name of
    the list element in square brackets, e.g., `@permitted [mode]`.
- `@default`: Default value for the argument. By default the default value
  from the function formals is displayed. This can be overwritten by using
  the `@default` tag.

- `@examplesx`: This tag can be used to mark the beginning of the examples
   section but doesn't affect the output, i.e., it can be omitted.

- `@caption`: Caption for the example. The caption is displayed as a subsection
  in the examples section. The caption can be followed by an arbitrary number
  of `@info` and `@code` tags.
- `@info`: Description of the example.
- `@code`: Code of the example.

     By default, any warning or error issued by the example code causes the
     building of the documentation to fail. If this is expected, the condition
     can be added to the `expected_cnds` option of the `@code` tag. E.g.,
     ```
     @code [expected_cnds = "warning"]
     ```

To use the roclet call `roxygen2::roxygenise(roclets =
"admiral::rdx_roclet")` or add to the `DESCRIPTION` file:
```
Roxygen: list(markdown = TRUE, roclets = c("collate", "namespace", "admiraldev::rdx_roclet"))
```

For more information on roxygen2 roclets see the [Extending
roxygen2](https://roxygen2.r-lib.org/articles/extending.html).
}
\keyword{documentation}
