% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quo.R
\name{replace_values_by_names}
\alias{replace_values_by_names}
\title{Replace Expression Value with Name}
\usage{
replace_values_by_names(expressions)
}
\arguments{
\item{expressions}{A list of expressions

\describe{
\item{Default value}{none}
}}
}
\value{
A list of expressions
}
\description{
Replace Expression Value with Name
}
\examples{
library(rlang)
replace_values_by_names(exprs(AVAL, ADT = convert_dtc_to_dt(EXSTDTC)))
}
\seealso{
Helpers for working with Quosures:
\code{\link[=add_suffix_to_vars]{add_suffix_to_vars()}},
\code{\link[=replace_symbol_in_expr]{replace_symbol_in_expr()}}
}
\concept{quo}
\keyword{quo}
