% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_helpers.R
\name{roxygen_float_comparison}
\alias{roxygen_float_comparison}
\title{Standard Text for Floating Point Comparisons in Conditions}
\usage{
roxygen_float_comparison()
}
\value{
A character string with the standardized documentation text about
floating point comparisons in conditions
}
\description{
This function provides standardized documentation text about comparing
floating point numbers in conditions, e.g., \code{PCHG <= -90}.
}
\details{
The benefits of having a programmatic way to write documentation is that if
any changes need to be made, modifying this function scales across the
codebase, can be tested, and is less prone to user-error such as typos or
grammar mistakes.
}
\examples{
roxygen_float_comparison()
}
\seealso{
Other documentation:
\code{\link[=roxygen_order_na_handling]{roxygen_order_na_handling()}},
\code{\link[=roxygen_param_by_vars]{roxygen_param_by_vars()}},
\code{\link[=roxygen_param_dataset]{roxygen_param_dataset()}},
\code{\link[=roxygen_save_memory]{roxygen_save_memory()}}
}
\concept{documentation}
\keyword{documentation}
