% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_helpers.R
\name{roxygen_param_by_vars}
\alias{roxygen_param_by_vars}
\title{Standard Text for the \code{by_vars} Argument}
\usage{
roxygen_param_by_vars(rename = FALSE)
}
\arguments{
\item{rename}{Should the renaming feature be documented?

If \code{TRUE}, the text includes information about renaming variables using
named elements in \code{by_vars}.

\describe{
\item{Permitted values}{\code{TRUE}, \code{FALSE}}
\item{Default value}{\code{FALSE}}
}}
}
\value{
A character string with the standardized documentation text for
the \code{by_vars} argument
}
\description{
This function provides standardized documentation text for the \code{by_vars}
argument used in \code{{admiral}} functions.
}
\details{
The benefits of having a programmatic way to write documentation is that if
any changes need to be made, modifying this function scales across the
codebase, can be tested, and is less prone to user-error such as typos or
grammar mistakes.
}
\examples{
roxygen_param_by_vars()

roxygen_param_by_vars(rename = TRUE)
}
\seealso{
Other documentation:
\code{\link[=roxygen_float_comparison]{roxygen_float_comparison()}},
\code{\link[=roxygen_order_na_handling]{roxygen_order_na_handling()}},
\code{\link[=roxygen_param_dataset]{roxygen_param_dataset()}},
\code{\link[=roxygen_save_memory]{roxygen_save_memory()}}
}
\concept{documentation}
\keyword{documentation}
