% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_helpers.R
\name{roxygen_save_memory}
\alias{roxygen_save_memory}
\title{Standard Note for Memory-Intensive Functions}
\usage{
roxygen_save_memory()
}
\value{
A character string with the standardized note about memory usage
}
\description{
This function provides a standardized note about memory consumption for
functions that create large temporary datasets.
}
\details{
The benefits of having a programmatic way to write documentation is that if
any changes need to be made, modifying this function scales across the
codebase, can be tested, and is less prone to user-error such as typos or
grammar mistakes.
}
\examples{
roxygen_save_memory()
}
\seealso{
Other documentation:
\code{\link[=roxygen_float_comparison]{roxygen_float_comparison()}},
\code{\link[=roxygen_order_na_handling]{roxygen_order_na_handling()}},
\code{\link[=roxygen_param_by_vars]{roxygen_param_by_vars()}},
\code{\link[=roxygen_param_dataset]{roxygen_param_dataset()}}
}
\concept{documentation}
\keyword{documentation}
