% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_rdx.R
\name{transform_examplesx}
\alias{transform_examplesx}
\title{Transform `@caption`, `@info` and `@code` tags into `@examplex` tags}
\usage{
transform_examplesx(block)
}
\arguments{
\item{block}{A block of roxygen tags of one function}
}
\value{
The block with the transformed tags
}
\description{
Transform `@caption`, `@info` and `@code` tags into `@examplex` tags
}
\keyword{internal}
