% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_rdx.R
\name{transform_param}
\alias{transform_param}
\title{Process `@permitted` and `@default` tags}
\usage{
transform_param(block, rdx_permitted_values)
}
\arguments{
\item{block}{A block of roxygen tags of one function}

\item{rdx_permitted_values}{A list of permitted values

  The name of the list item can be referenced in the `@permitted` tag by
  specifying the name in square brackets, e.g., `@permitted [mode]`.}
}
\value{
The block with the transformed tags
}
\description{
This function processes the `@permitted` and `@default` tags and moves them
into the `@param` tag. The default value is taken from the function call if
not specified.
}
\keyword{internal}
