% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{vars2chr}
\alias{vars2chr}
\title{Turn a List of Expressions into a Character Vector}
\usage{
vars2chr(expressions)
}
\arguments{
\item{expressions}{A `list` of expressions created using `exprs()`}
}
\value{
A character vector
}
\description{
Turn a List of Expressions into a Character Vector
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(rlang)

vars2chr(exprs(USUBJID, AVAL))
}
\seealso{
Developer Utility Functions:
\code{\link{\%notin\%}()},
\code{\link{contains_vars}()},
\code{\link{convert_dtm_to_dtc}()},
\code{\link{extract_vars}()},
\code{\link{filter_if}()}
}
\concept{dev_utility}
\keyword{dev_utility}
