% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters.R
\name{admiral_linters}
\alias{admiral_linters}
\title{Linter configurations for \code{{admiral}} family of packages}
\usage{
admiral_linters()
}
\description{
This function lists out the linter configurations for the \code{{admiral}}
family of packages..
}
\details{
Configurations of special note:
\itemize{
\item Enforcing use of \code{cli::cli_abort()} over \code{stop()}
\item Enforcing use of \code{cli::cli_warn()} over \code{warning()}
}
}
\keyword{internal}
