% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiraldev-package.R
\docType{package}
\name{admiraldev-package}
\alias{admiraldev}
\alias{admiraldev-package}
\title{admiraldev: Utility Functions and Development Tools for the Admiral Package Family}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utility functions to check data, variables and conditions for functions used in 'admiral' and 'admiral' extension packages. Additional utility helper functions to assist developers with maintaining documentation, testing and general upkeep of 'admiral' and 'admiral' extension packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiraldev/}
  \item \url{https://github.com/pharmaverse/admiraldev/}
  \item Report bugs at \url{https://github.com/pharmaverse/admiraldev/issues}
}

}
\author{
\strong{Maintainer}: Ben Straub \email{ben.x.straub@gsk.com}

Authors:
\itemize{
  \item Ben Straub \email{ben.x.straub@gsk.com}
  \item Stefan Bundfuss (\href{https://orcid.org/0009-0005-0027-1198}{ORCID})
  \item Arianna Cascone (\href{https://orcid.org/0000-0001-5948-2831}{ORCID})
  \item Kristin Dahnert
  \item Jeffrey Dickinson
  \item Ross Farrugia
  \item Fanny Gautier
  \item Edoardo Mancini (\href{https://orcid.org/0009-0006-4899-8641}{ORCID})
  \item Gordon Miller
  \item Lina Patil
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
}

}
\keyword{internal}
