% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{arg_name}
\alias{arg_name}
\title{Extract Argument Name from an Expression}
\usage{
arg_name(expr)
}
\arguments{
\item{expr}{An expression created inside a function using `substitute()`}
}
\value{
`character` vector
}
\description{
`r lifecycle::badge("deprecated")`

This function is *deprecated*, please use `rlang::caller_arg()` instead.
}
\seealso{
Other deprecated: 
\code{\link{\%or\%}()},
\code{\link{enumerate}()},
\code{\link{friendly_type_of}()},
\code{\link{valid_time_units}()},
\code{\link{what_is_it}()}
}
\concept{deprecated}
\keyword{deprecated}
