% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{arg_name}
\alias{arg_name}
\title{Extract Argument Name from an Expression}
\usage{
arg_name(expr)
}
\arguments{
\item{expr}{An expression created inside a function using \code{substitute()}

\describe{
\item{Default value}{none}
}}
}
\value{
\code{character} vector
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please use \code{rlang::caller_arg()} instead.
}
\seealso{
Other deprecated:
\code{\link{\%or\%}},
\code{\link[=enumerate]{enumerate()}},
\code{\link[=friendly_type_of]{friendly_type_of()}},
\code{\link[=valid_time_units]{valid_time_units()}},
\code{\link[=what_is_it]{what_is_it()}}
}
\concept{deprecated}
\keyword{deprecated}
