% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_expr}
\alias{assert_expr}
\title{Assert Argument is an Expression}
\usage{
assert_expr(
  arg,
  optional = FALSE,
  arg_name = gsub("^enexpr\\\\((.*)\\\\)$", "\\\\1", rlang::caller_arg(arg)),
  message = NULL,
  class = "assert_expr",
  call = parent.frame()
)
}
\arguments{
\item{arg}{A function argument to be checked}

\item{optional}{Is the checked argument optional? If set to `FALSE` and `arg`
is `NULL` then an error is thrown}

\item{arg_name}{By default the expression specified for `arg` is used. If it is
of the form `enexpr(<argument name>)`, the `enexpr()` part is removed. For example
if `arg = enexpr(filter_add)` is specified, `arg_name` defaults to `"filter_add"`}

\item{message}{string passed to `cli::cli_abort(message)`.
When `NULL`, default messaging is used (see examples for default messages).
`"{arg_name}"` can be used in messaging.}

\item{class}{Subclass of the condition.}

\item{call}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}
}
\value{
The function throws an error if `arg` is not an expression, i.e. either
a symbol or a call, or returns the input invisibly otherwise
}
\description{
Assert Argument is an Expression
}
\seealso{
Checks for valid input and returns warning or errors messages:
\code{\link{assert_atomic_vector}()},
\code{\link{assert_character_scalar}()},
\code{\link{assert_character_vector}()},
\code{\link{assert_data_frame}()},
\code{\link{assert_date_vector}()},
\code{\link{assert_expr_list}()},
\code{\link{assert_filter_cond}()},
\code{\link{assert_function}()},
\code{\link{assert_integer_scalar}()},
\code{\link{assert_list_element}()},
\code{\link{assert_list_of}()},
\code{\link{assert_logical_scalar}()},
\code{\link{assert_named}()},
\code{\link{assert_numeric_vector}()},
\code{\link{assert_one_to_one}()},
\code{\link{assert_param_does_not_exist}()},
\code{\link{assert_s3_class}()},
\code{\link{assert_same_type}()},
\code{\link{assert_symbol}()},
\code{\link{assert_unit}()},
\code{\link{assert_vars}()},
\code{\link{assert_varval_list}()}
}
\concept{assertion}
\keyword{assertion}
