% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_rdx.R
\name{capture_message}
\alias{capture_message}
\title{Captures Messages Preserving the Message Redirection}
\usage{
capture_message(expr)
}
\arguments{
\item{expr}{An R expression to evaluate

\describe{
\item{Default value}{none}
}}
}
\value{
A character vector of captured messages
}
\description{
Captures Messages Preserving the Message Redirection
}
\keyword{internal}
