% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_rdx.R
\name{capture_output}
\alias{capture_output}
\title{Capture Output and Messages}
\usage{
capture_output(expr, expected_cnds = NULL, env = caller_env())
}
\arguments{
\item{expr}{An R expression to evaluate

\describe{
\item{Default value}{none}
}}

\item{expected_cnds}{A character vector of expected conditions

If the expression issues a condition of a class that is in this vector, the
condition is ignored but added to the return value.

Otherwise, an error is issued.

\describe{
\item{Default value}{\code{NULL}}
}}

\item{env}{The environment in which to evaluate the expression

\describe{
\item{Default value}{\code{caller_env()}}
}}
}
\value{
A character vector of captured output and messages
}
\description{
The function captures both output and expected messages from an R expression.
If the expression results in an unexpected message, an error is issued.
}
