% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{convert_dtm_to_dtc}
\alias{convert_dtm_to_dtc}
\title{Helper Function to Convert Date (or Date-time) Objects to Characters of dtc Format
(-DTC type of variable)}
\usage{
convert_dtm_to_dtc(dtm)
}
\arguments{
\item{dtm}{date or date-time

\describe{
\item{Default value}{none}
}}
}
\value{
\code{character} vector
}
\description{
Helper Function to Convert Date (or Date-time) Objects to Characters of dtc Format
(-DTC type of variable)
}
\seealso{
Developer Utility Functions:
\code{\link{\%notin\%}},
\code{\link[=contains_vars]{contains_vars()}},
\code{\link[=extract_vars]{extract_vars()}},
\code{\link[=filter_if]{filter_if()}},
\code{\link[=vars2chr]{vars2chr()}}
}
\concept{dev_utility}
\keyword{dev_utility}
