% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle_admiral.R
\name{deprecate_inform}
\alias{deprecate_inform}
\title{Deprecation with Soft Message}
\usage{
deprecate_inform(...)
}
\arguments{
\item{...}{See documentation for \code{lifecycle::deprecate_soft()} for argument use.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Wrapper around \code{lifecycle::deprecate_soft()}.
}
\examples{
# A Phase 1 deprecated function with custom bulleted list:
deprecate_inform(
  when = "1.0.0",
  what = "foo()",
  details = c(
    #     'x = "This message will turn into a warning with release of x.y.z",
    i = "See admiral's deprecation guidance:
https://pharmaverse.github.io/admiraldev/dev/articles/programming_strategy.html#deprecation"
  )
)

}
\concept{messages}
\keyword{messages}
