% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle_admiral.R
\name{deprecate_inform}
\alias{deprecate_inform}
\title{Deprecation with Soft Message}
\usage{
deprecate_inform(...)
}
\arguments{
\item{...}{}
}
\value{
\code{NULL}, invisibly.
}
\description{
Wrapper around \code{lifecycle::deprecate_soft()}.
See documentation for \code{lifecycle::deprecate_soft()} for argument use.
}
\examples{
# A Phase 1 deprecated function with custom bulleted list:
deprecate_warn(
  when = "1.0.0",
  what = "foo()",
  details = c(
    i = "See admiral's deprecation guidance:
https://pharmaverse.github.io/admiraldev/dev/articles/programming_strategy.html#deprecation"
)
)



}
\concept{messages}
\keyword{messages}
