% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\name{enumerate}
\alias{enumerate}
\title{Enumerate Multiple Elements}
\usage{
enumerate(x, quote_fun = backquote, conjunction = "and")
}
\arguments{
\item{x}{A vector or list}

\item{quote_fun}{Quoting function, defaults to `backquote`. If set to `NULL`,
the elements are not quoted.}

\item{conjunction}{Character to be used in the message, defaults to `"and"`.}
}
\value{
A `character` vector
}
\description{
`r lifecycle::badge("deprecated")`

This function is *deprecated*, please use `cli` functionality instead.
}
\seealso{
Other deprecated: 
\code{\link{\%or\%}()},
\code{\link{arg_name}()},
\code{\link{friendly_type_of}()},
\code{\link{valid_time_units}()},
\code{\link{what_is_it}()}
}
\concept{deprecated}
\keyword{deprecated}
