% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\name{enumerate}
\alias{enumerate}
\title{Enumerate Multiple Elements}
\usage{
enumerate(x, quote_fun = backquote, conjunction = "and")
}
\arguments{
\item{x}{A vector or list

\describe{
\item{Default value}{none}
}}

\item{quote_fun}{Quoting function, defaults to \code{backquote}. If set to \code{NULL},
the elements are not quoted.

\describe{
\item{Default value}{\code{backquote}}
}}

\item{conjunction}{Character to be used in the message, defaults to \code{"and"}.

\describe{
\item{Default value}{\code{"and"}}
}}
}
\value{
A \code{character} vector
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please use \code{cli} functionality instead.
}
\seealso{
Other deprecated: 
\code{\link{\%or\%}()},
\code{\link{arg_name}()},
\code{\link{friendly_type_of}()},
\code{\link{valid_time_units}()},
\code{\link{what_is_it}()}
}
\concept{deprecated}
\keyword{deprecated}
