% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\name{enumerate}
\alias{enumerate}
\title{Enumerate Multiple Elements}
\usage{
enumerate(x, quote_fun = backquote, conjunction = "and")
}
\arguments{
\item{x}{A vector or list

\describe{
\item{Default value}{none}
}}

\item{quote_fun}{Quoting function, defaults to \code{backquote}. If set to \code{NULL},
the elements are not quoted.

\describe{
\item{Default value}{\code{backquote}}
}}

\item{conjunction}{Character to be used in the message, defaults to \code{"and"}.

\describe{
\item{Default value}{\code{"and"}}
}}
}
\value{
A \code{character} vector
}
\description{
Enumerate multiple elements of a vector or list.
}
\examples{
enumerate(c("one", "two", "three"))

enumerate(c(1, 2, 3), quote_fun = NULL)

}
\seealso{
Helpers for working with Quotes and Quoting: 
\code{\link{backquote}()},
\code{\link{dquote}()},
\code{\link{squote}()}
}
\concept{quote}
\keyword{quote}
