% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_rdx.R
\name{execute_example}
\alias{execute_example}
\title{Execute Example Code}
\usage{
execute_example(code, expected_cnds = NULL, env = caller_env())
}
\arguments{
\item{code}{The example code

  The code is expected to be a character vector of R code lines.}

\item{expected_cnds}{Expected conditions}

\item{env}{The environment in which to evaluate the example code}
}
\value{
A character vector of the input code and the output and messages
  created by the code. Output and messages are prefixed by `"#>"`.
}
\description{
This function executes the example code and captures the output and messages.
If the example code issues an unexpected message, an error is issued.
}
\keyword{internal}
