% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quo.R
\name{expr_c}
\alias{expr_c}
\title{Concatenate One or More Expressions}
\usage{
expr_c(...)
}
\arguments{
\item{...}{One or more expressions or list of expressions}
}
\value{
A list of expressions
}
\description{
Concatenate One or More Expressions
}
\seealso{
Helpers for working with Quosures:
\code{\link{add_suffix_to_vars}()},
\code{\link{replace_symbol_in_expr}()},
\code{\link{replace_values_by_names}()}
}
\concept{quo}
\keyword{quo}
