% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{extract_vars}
\alias{extract_vars}
\title{Extract All Symbols from a List of Expressions}
\usage{
extract_vars(x, side = "lhs")
}
\arguments{
\item{x}{An \code{R} object

\describe{
\item{Default value}{none}
}}

\item{side}{One of \code{"lhs"} (the default) or \code{"rhs"} for formulas

\describe{
\item{Default value}{\code{"lhs"}}
}}
}
\value{
A list of expressions
}
\description{
Extract All Symbols from a List of Expressions
}
\examples{
library(rlang)
extract_vars(exprs(PARAMCD, (BASE - AVAL) / BASE + 100))
extract_vars(AVAL ~ ARMCD + AGEGR1)
extract_vars(AVAL ~ ARMCD + AGEGR1, side = "rhs")
}
\seealso{
Developer Utility Functions:
\code{\link{\%notin\%}},
\code{\link[=contains_vars]{contains_vars()}},
\code{\link[=convert_dtm_to_dtc]{convert_dtm_to_dtc()}},
\code{\link[=filter_if]{filter_if()}},
\code{\link[=vars2chr]{vars2chr()}}
}
\concept{dev_utility}
\keyword{dev_utility}
