% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{filter_if}
\alias{filter_if}
\title{Optional Filter}
\usage{
filter_if(dataset, filter)
}
\arguments{
\item{dataset}{Input dataset}

\item{filter}{A filter condition. Must be an expression.}
}
\value{
A `data.frame` containing all rows in `dataset` matching `filter` or
just `dataset` if `filter` is `NULL`
}
\description{
Filters the input dataset if the provided expression is not `NULL`
}
\seealso{
Developer Utility Functions:
\code{\link{\%notin\%}()},
\code{\link{contains_vars}()},
\code{\link{convert_dtm_to_dtc}()},
\code{\link{extract_vars}()},
\code{\link{vars2chr}()}
}
\concept{dev_utility}
\keyword{dev_utility}
