% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_friendly_type.R
\name{friendly_type_of}
\alias{friendly_type_of}
\title{Return English-friendly messaging for object-types}
\usage{
friendly_type_of(x, value = TRUE, length = FALSE)
}
\arguments{
\item{x}{Any R object.}

\item{value}{Whether to describe the value of `x`.}

\item{length}{Whether to mention the length of vectors and lists.}
}
\value{
A string describing the type. Starts with an indefinite
  article, e.g. "an integer vector".
}
\description{
`r lifecycle::badge("deprecated")`

This function is *deprecated*, please use `cli` functionality instead.
}
\details{
This helper function aids us in forming user-friendly messages that gets
called through `what_is_it()`, which is often used in the assertion functions
to identify what object-type the user passed through an argument instead of
an expected-type.
}
\seealso{
Other deprecated: 
\code{\link{\%or\%}()},
\code{\link{arg_name}()},
\code{\link{enumerate}()},
\code{\link{valid_time_units}()},
\code{\link{what_is_it}()}
}
\concept{deprecated}
\keyword{deprecated}
