% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Retrieve a Dataset from the \code{admiraldev_environment} environment}
\usage{
get_dataset(name)
}
\arguments{
\item{name}{The name of the dataset to retrieve}
}
\value{
A \code{data.frame}
}
\description{
Retrieve a Dataset from the \code{admiraldev_environment} environment
}
\details{
Once in a while, we may encounter "locked binding for 'xxx'." errors
during the development process while building out functions. This may arise because
we want to create dynamic data/objects based on user-inputs that need modification
at points in time after the package has been loaded. To manage such data or objects,
R has a data structure known as an 'environment'. These environment objects are created
at build time, but can be populated with values after the package has been loaded and
update those values over the course of an R session.

Currently we only support two datasets inside our \code{admiraldev_environment} object:
\itemize{
\item \code{one_to_many}
\item \code{many_to_one}
}

The purpose of \code{get_dataset} is to retrieve the datasets contained
inside \code{admiraldev_environment}.
}
\concept{datasets}
\keyword{datasets}
