% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_duplicates}
\alias{get_duplicates}
\title{Get Duplicates From a Vector}
\usage{
get_duplicates(x)
}
\arguments{
\item{x}{An atomic vector}
}
\value{
A vector of the same type as `x` contain duplicate values
}
\description{
Get Duplicates From a Vector
}
\examples{
get_duplicates(1:10)

get_duplicates(c("a", "a", "b", "c", "d", "d"))
}
\seealso{
Brings something to you!?!:
\code{\link{get_constant_vars}()},
\code{\link{get_dataset}()},
\code{\link{get_source_vars}()}
}
\concept{get}
\keyword{get}
