% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_rdx.R
\name{get_param_tag}
\alias{get_param_tag}
\title{Get the content of the \verb{@param} tag}
\usage{
get_param_tag(act_param, defaults)
}
\arguments{
\item{act_param}{The content of the argument description to format

A list with three named elements is expected:
\itemize{
\item \code{tag}: The \verb{@param} tag
\item \code{permitted}: The permitted values
\item \code{default}: The default value
}

\describe{
\item{Default value}{none}
}}

\item{defaults}{A list of default values from the function call

\describe{
\item{Default value}{none}
}}
}
\value{
The formatted \verb{@param} tag
}
\description{
This function adds the permitted values and the default value to the
description of the \verb{@param} tag. They are added as a definition list
(\verb{\\describe\{\}}).
}
\keyword{internal}
