% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_source_vars}
\alias{get_source_vars}
\title{Get Source Variables from a List of Expressions}
\usage{
get_source_vars(expressions)
}
\arguments{
\item{expressions}{A list of expressions}
}
\value{
A list of expressions
}
\description{
Get Source Variables from a List of Expressions
}
\seealso{
Brings something to you!?!: 
\code{\link{get_constant_vars}()},
\code{\link{get_dataset}()},
\code{\link{get_duplicates}()}
}
\concept{get}
\keyword{get}
