% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negated Value Matching}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{The values to be matched

\describe{
\item{Default value}{none}
}}

\item{table}{The values to be matched against

\describe{
\item{Default value}{none}
}}
}
\value{
A \code{logical} vector
}
\description{
Returns a \code{logical} vector indicating if there is \emph{no} match of the
left operand in the right operand.
}
\seealso{
Developer Utility Functions:
\code{\link[=contains_vars]{contains_vars()}},
\code{\link[=convert_dtm_to_dtc]{convert_dtm_to_dtc()}},
\code{\link[=extract_vars]{extract_vars()}},
\code{\link[=filter_if]{filter_if()}},
\code{\link[=vars2chr]{vars2chr()}}
}
\concept{dev_utility}
\keyword{dev_utility}
