% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{\%or\%}
\alias{\%or\%}
\title{Or}
\usage{
lhs \%or\% rhs
}
\arguments{
\item{lhs}{Any valid R expression}

\item{rhs}{Any valid R expression}
}
\value{
Either the result of evaluating `lhs`, `rhs` or an error
}
\description{
`r lifecycle::badge("deprecated")`

This function is *deprecated*. Please get in touch if you are using this function!
}
\details{
The function evaluates the expression `lhs` and if this expression results
in an error, it catches that error and proceeds with evaluating the expression
`rhs` and returns that result.
}
\seealso{
Other deprecated: 
\code{\link{arg_name}()},
\code{\link{enumerate}()},
\code{\link{friendly_type_of}()},
\code{\link{valid_time_units}()},
\code{\link{what_is_it}()}
}
\concept{deprecated}
\keyword{deprecated}
