% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_order_vars}
\alias{is_order_vars}
\title{Is order vars?}
\usage{
is_order_vars(arg)
}
\arguments{
\item{arg}{An R object}
}
\value{
`FALSE` if the argument is not a list of order vars
}
\description{
Check if inputs are created using `exprs()` or calls involving `desc()`
}
\seealso{
Identifies type of Object with return of TRUE/FALSE:
\code{\link{is_auto}()},
\code{\link{is_valid_dtc}()}
}
\concept{is}
\keyword{is}
