% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_valid_dtc}
\alias{is_valid_dtc}
\title{Is this string a valid DTC?}
\usage{
is_valid_dtc(arg, valid_full_date = FALSE, valid_full_datetime = FALSE)
}
\arguments{
\item{arg}{A \code{character} vector

\describe{
\item{Default value}{none}
}}

\item{valid_full_date}{If \code{TRUE}, the function will check if the string is a
full date (i.e. it is exactly of the form \code{"yyyy-mm-dd"} and can be parsed by
\code{ymd()}).

\describe{
\item{Permitted values}{\code{TRUE}, \code{FALSE}}
\item{Default value}{\code{FALSE}}
}}

\item{valid_full_datetime}{If \code{TRUE}, the function will check if the string is a
full datetime (i.e. it is exactly of the form \code{"yyyy-mm-ddThh:mm:ss"} and can be
parsed by \code{ymd_hms()}).

\describe{
\item{Permitted values}{\code{TRUE}, \code{FALSE}}
\item{Default value}{\code{FALSE}}
}}
}
\value{
\code{TRUE} if the argument is a valid \code{--DTC} string, \code{FALSE} otherwise
}
\description{
Is this string a valid DTC?
}
\details{
If both \code{check_full_date} and \code{check_full_time} are set to \code{FALSE}, the
function will only check if the string is a valid DTC string, which can also be a
partial date (e.g. \code{"2024"} or \code{"2024-05"}) or a partial datetime (e.g.
\code{"2024-05-01T12"}).
}
\seealso{
Identifies type of Object with return of TRUE/FALSE:
\code{\link[=is_auto]{is_auto()}},
\code{\link[=is_order_vars]{is_order_vars()}}
}
\concept{is}
\keyword{is}
