% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_helpers.R
\name{roxygen_order_na_handling}
\alias{roxygen_order_na_handling}
\title{Standard Text for NA Handling in Sorting Variables}
\usage{
roxygen_order_na_handling()
}
\value{
A character string with the standardized documentation text about
NA handling in sort order
}
\description{
This function provides standardized documentation text about the handling
of \code{NA} values in sorting variables.
}
\details{
The benefits of having a programmatic way to write documentation is that if
any changes need to be made, modifying this function scales across the
codebase, can be tested, and is less prone to user-error such as typos or
grammar mistakes.
}
\examples{
roxygen_order_na_handling()
}
\seealso{
Other documentation: 
\code{\link{roxygen_param_by_vars}()},
\code{\link{roxygen_param_dataset}()},
\code{\link{roxygen_save_memory}()}
}
\concept{documentation}
\keyword{documentation}
