% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_helpers.R
\name{roxygen_param_dataset}
\alias{roxygen_param_dataset}
\title{Standard Text for the `dataset` Argument}
\usage{
roxygen_param_dataset(expected_vars = NULL)
}
\arguments{
\item{expected_vars}{Variables expected in the dataset

  A character vector of variable names expected to be present in the
  `dataset` argument.}
}
\value{
A character string with the standardized documentation text for
  the `dataset` argument
}
\description{
This function provides standardized documentation text for the `dataset`
argument used in `{admiral}` functions.
}
\details{
The benefits of having a programmatic way to write documentation is that if
any changes need to be made, modifying this function scales across the
codebase, can be tested, and is less prone to user-error such as typos or
grammar mistakes.
}
\examples{
roxygen_param_dataset()

roxygen_param_dataset(expected_vars = c("by_vars"))
}
\seealso{
Other documentation: 
\code{\link{roxygen_order_na_handling}()},
\code{\link{roxygen_param_by_vars}()},
\code{\link{roxygen_save_memory}()}
}
\concept{documentation}
\keyword{documentation}
