% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\name{squote}
\alias{squote}
\title{Wrap a String in Single Quotes}
\usage{
squote(x)
}
\arguments{
\item{x}{A `character` vector}
}
\value{
A `character` vector
}
\description{
Wrap a String in Single Quotes
}
\seealso{
Helpers for working with Quotes and Quoting:
\code{\link{backquote}()},
\code{\link{dquote}()}
}
\concept{quote}
\keyword{quote}
