% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_rdx.R
\name{transform_examplesx}
\alias{transform_examplesx}
\title{Transform \verb{@caption}, \verb{@info} and \verb{@code} tags into \verb{@examplex} tags}
\usage{
transform_examplesx(block)
}
\arguments{
\item{block}{A block of roxygen tags of one function

\describe{
\item{Default value}{none}
}}
}
\value{
The block with the transformed tags
}
\description{
Transform \verb{@caption}, \verb{@info} and \verb{@code} tags into \verb{@examplex} tags
}
\keyword{internal}
