% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{valid_time_units}
\alias{valid_time_units}
\title{Valid Time Units}
\usage{
valid_time_units()
}
\value{
A \code{character} vector of valid time units
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}. Please get in touch if you are using this function!

Contains the acceptable character vector of valid time units
}
\seealso{
Other deprecated:
\code{\link{\%or\%}},
\code{\link[=arg_name]{arg_name()}},
\code{\link[=enumerate]{enumerate()}},
\code{\link[=friendly_type_of]{friendly_type_of()}},
\code{\link[=what_is_it]{what_is_it()}}
}
\concept{deprecated}
\keyword{deprecated}
