% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{valid_time_units}
\alias{valid_time_units}
\title{Valid Time Units}
\usage{
valid_time_units()
}
\value{
A `character` vector of valid time units
}
\description{
`r lifecycle::badge("deprecated")`

This function is *deprecated*. Please get in touch if you are using this function!

Contains the acceptable character vector of valid time units
}
\seealso{
Other deprecated: 
\code{\link{\%or\%}()},
\code{\link{arg_name}()},
\code{\link{enumerate}()},
\code{\link{friendly_type_of}()},
\code{\link{what_is_it}()}
}
\concept{deprecated}
\keyword{deprecated}
