% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what.R
\name{what_is_it}
\alias{what_is_it}
\title{What Kind of Object is This?}
\usage{
what_is_it(x)
}
\arguments{
\item{x}{Any R object

\describe{
\item{Default value}{none}
}}
}
\value{
A \code{character} description of the type of \code{x}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please use \code{cli} functionality instead.
}
\seealso{
Other deprecated: 
\code{\link{\%or\%}()},
\code{\link{arg_name}()},
\code{\link{enumerate}()},
\code{\link{friendly_type_of}()},
\code{\link{valid_time_units}()}
}
\concept{deprecated}
\keyword{deprecated}
